convert to isometric

+++ fdf.h
@@ -26,8 +26,8 @@
 
 typedef struct s_brsh
 {
-	int	x0;
-	int	x1;
+	double	x0;
+	double	x1;
 	int	y0;
 	int	y1;
 }	t_brsh;
